refactor: Move pass-though specific code to its own package
