+++ src/main/java/org/opentripplanner/routing/algorithm/transferoptimization/model/passthrough/PassThroughFilterFactory.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.transferoptimization.services;
+package org.opentripplanner.routing.algorithm.transferoptimization.model.passthrough;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,10 +7,10 @@ import org.opentripplanner.raptor.api.request.PassThroughPoint;
 import org.opentripplanner.raptor.path.PathBuilderLeg;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.MinCostPathTailFilter;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.OptimizedPathTail;
-import org.opentripplanner.routing.algorithm.transferoptimization.model.PassThroughPathTailFilter;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.PathTailFilter;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.PathTailFilterFactory;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.TripToTripTransfer;
+import org.opentripplanner.routing.algorithm.transferoptimization.services.TransferOptimizedFilterFactory;
 
 public class PassThroughFilterFactory<T extends RaptorTripSchedule>
   implements PathTailFilterFactory<T> {
